Немного рефакторит
